pyproject.toml
[tool.poetry]
name = "drive-gpt"
version = "0.1.0"
description = ""
authors = ["Sean Muirhead <smmuirhead100@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.51.2"
pydantic = "^2.9.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
